refactor: 추상화 단계 수정
